SounderRecorder: can not save file on new android version
because of _data column is deprecated in the Mediastore,
and there are AOB issues in the code.

SounderRecorder app can not save file
based on new android version.

We will offer a solution by lowering down
the SDK version and fix dabase access errors due to
the new restriction on platform upgrading.

Change-Id: I1f0b38261db49f04c31157128da4a7a5871c48b0
CRs-Fixed: 2372431